requirements.txt
APScheduler==3.9.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-filter==21.1
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
idna==3.3
importlib-resources==5.4.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
zipp==3.6.0
